# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
...
  internal_features_teams_transport_http.CreateTeamRequest:
    properties:
      contact_link:
        maxLength: 2000
        type: string
      description:
        maxLength: 1000
        type: string
      desired_rating:
        maxLength: 200
        type: string
      game_id:
        type: integer
      is_rating_required:
        type: boolean
      platform_id:
        type: integer
      slots_total:
        type: integer
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - title
    type: object
  internal_features_teams_transport_http.CreateTeamResponse:
    properties:
      contact_link:
        type: string
      created_at:
        type: string
      description:
        type: string
      desired_rating:
        type: string
      game_id:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      is_rating_required:
        type: boolean
      members:
        items:
          $ref: '#/definitions/internal_features_teams_transport_http.TeamMemberDTOResponse'
        type: array
      owner_id:
        type: string
      platform_id:
        type: integer
      slots_taken:
        type: integer
      slots_total:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
...
  internal_features_teams_transport_http.TeamDTOResponse:
    properties:
      contact_link:
        type: string
      created_at:
        type: string
      description:
        type: string
      desired_rating:
        type: string
      game_id:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      is_rating_required:
        type: boolean
      members:
        items:
          $ref: '#/definitions/internal_features_teams_transport_http.TeamMemberDTOResponse'
        type: array
      owner_id:
        type: string
      platform_id:
        type: integer
      slots_taken:
        type: integer
      slots_total:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_teams_transport_http.TeamMemberDTOResponse:
    properties:
      nickname:
        type: string
      user_id:
        type: string
    type: object
...
  internal_features_users_transport_http.PatchProfileRequest:
    properties:
      avatar_key:
        $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string'
      bio:
        $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string'
      nickname:
        $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_types.Nullable-string'
    type: object
  internal_features_users_transport_http.PatchProfileResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      id:
        type: string
      nickname:
        type: string
      version:
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: DPS Backend REST-API Schema
  title: DPS Backend API
  version: "1.0"
paths:
...
  /teams:
    get:
      description: Returns active teams with optional filtering by search query, game
        and platform.
      parameters:
      - description: Search by team title
        in: query
        name: q
        type: string
      - description: Filter by game ID
        in: query
        name: game_id
        type: integer
      - description: Filter by platform ID
        in: query
        name: platform_id
        type: integer
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_features_teams_transport_http.TeamDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      summary: List teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Creates a team for the authenticated user as its owner. The owner
        automatically becomes a member.
      parameters:
      - description: Create team payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_teams_transport_http.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_features_teams_transport_http.CreateTeamResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create team
      tags:
      - teams
...
  /users/profile/me:
    delete:
      description: Deletes the authenticated user's profile (account). Only the owner
        can delete their own profile.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: 'Partially updates the authenticated user''s public profile (nickname,
        bio, avatar preset). A field is updated only if it is present in the body;
        pass "bio": null to clear it. Only the owner can modify their own profile.'
      parameters:
      - description: Profile patch payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchProfileResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict (optimistic lock)
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_M1sterZag_Dont_Play_Separately_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Введите access-токен в формате "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"